Move URL verification logic into its own file
